+++ src/main/java/modules/gestionFranquicias/services/FranquiciaService.java
@@ -18,7 +18,7 @@ public class FranquiciaService {
     @GET
     @Path("/{id}")
     @Produces("application/json")
-    public String get(@PathParam("id") int id) {
+    public String get(@PathParam("id") long id) {
         FranquiciaModel f = franquiciaController.find(id);
         Gson gson = new Gson();
         String json = gson.toJson(f);
@@ -29,7 +29,7 @@ public class FranquiciaService {
     @Path("/franquiciaActual")
     @Produces("application/json")
     public String get() {
-        Integer f = franquiciaController.find();
+        Long f = franquiciaController.find();
         Gson gson = new Gson();
         String json = gson.toJson(f);
         return json;
@@ -48,7 +48,7 @@ public class FranquiciaService {
     @GET
     @Path("proveedor/{id}")
     @Produces("application/json")
-    public String getAll(@PathParam("id") int idProveedor) {
+    public String getAll(@PathParam("id") long idProveedor) {
         List<FranquiciaModel> franquicias = franquiciaController.findAll(idProveedor);
         Gson gson = new Gson();
         String json = gson.toJson(franquicias);
@@ -67,7 +67,7 @@ public class FranquiciaService {
     @PUT
     @Path("/{id}")
     @Consumes("application/json")
-    public Response update(@PathParam("id") int id, String json) {
+    public Response update(@PathParam("id") long id, String json) {
         Gson gson = new Gson();
         FranquiciaModel f = gson.fromJson(json, FranquiciaModel.class);
         if (franquiciaController.update(id, f)) return javax.ws.rs.core.Response.accepted().build();
@@ -76,7 +76,7 @@ public class FranquiciaService {
 
     @DELETE
     @Path("/{id}")
-    public void remove(@PathParam("id") int id) {
+    public void remove(@PathParam("id") long id) {
         franquiciaController.remove(id);
     }
 }
